requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.7.0
click==8.1.7
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.1.0
importlib-metadata==6.8.0
isort==5.12.0
Markdown==3.4.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.16.2
